api_gateway/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
kafka-python==2.2.0
pydantic==2.5.3